ER Diagram for our SIS project
The Entity-Relationship Diagram (ERD) for the Student Information System (SIS) represents the system's data model and how different entities (tables) in the PostgreSQL database are related to each other. The diagram provides a high-level overview of the structure of the database and defines the relationships between various components, such as students, courses, grades, and registrations. Below is a description of the entities and their relationships. It's first look.

(chen notation)